requirements.txt
python-docx==1.0.1
requests==2.31.0

